🚨 [Tests]: runtime-configテストファイルを機能別に分割
- runtime-config.test.tsを3つのファイルに分割：
  - runtime-config.basic.test.ts (基本的な設定機能のテスト - 5テスト)
  - runtime-config.figma-info.test.ts (Figma情報管理のテスト - 3テスト)
  - runtime-config.figma-url.test.ts (URL解析機能のテスト - 2テスト)
- Kent C. Doddsの「Avoid nesting when you're testing」の原則に従いフラットなテスト構造を採用
- テストの可読性と保守性を向上
